+++ SubModules/Render/D3D11/D3D11Renderer.cpp
@@ -90,23 +90,6 @@ void RRenderer::BeginRender()
 	}
 }
 
-void RRenderer::Render()
-{
-	// Update Scene Resources
-	{
-		UpdateGlobalConstant();
-		// Light
-		m_resourceManager->UpdateStructuredBuffer(sizeof(RLightConstant), static_cast<UINT>(m_sceneLightsIndex), m_sceneLights, m_sceneLightsBuffer);
-	}
-
-	RenderSkybox();
-
-	RenderOpaques();
-
-	RenderTransparent();
-}
-
-
 
 void RRenderer::EndRender()
 {
@@ -227,57 +210,6 @@ void RRenderer::Compute(const RComputeShader* pComputeShader, const WCHAR* pTask
 	m_deviceResources->PIXEndEvent();
 }
 
-RTexture* RRenderer::CreateTexture3D(const UINT width, const UINT height, const UINT depth, const DXGI_FORMAT format)
-{
-	return m_resourceManager->CreateTexture3D(width, height, depth, format);
-}
-
-RTexture* RRenderer::CreateTexture2D(const UINT width, const UINT height, const UINT count, const DXGI_FORMAT format)
-{
-	return m_resourceManager->CreateTexture2D(width, height, count, format);
-}
-
-RTexture* RRenderer::CreateStructuredBuffer(const UINT uElementSize, const UINT uElementCount, const void* pInitData)
-{
-	return m_resourceManager->CreateStructuredBuffer(uElementSize, uElementCount, pInitData);
-}
-
-void RRenderer::SetCamera(const Camera* pCamera)
-{
-	MY_ASSERT(pCamera != nullptr);
-
-	m_camera = pCamera;
-}
-
-void RRenderer::SetSkybox(Skybox* pSkybox)
-{
-	MY_ASSERT(pSkybox != nullptr);
-	m_skybox = pSkybox;
-}
-
-void RRenderer::SetSunLight(const Light* pLight)
-{
-	MY_ASSERT(pLight != nullptr);
-	m_sunLight = pLight;
-}
-
-
-
-void RRenderer::AddLight(const Light* pLight)
-{
-	if (m_sceneLightsIndex >= MAX_SCENE_LIGHTS_COUNT)
-	{
-		return;
-	}
-
-	m_sceneLights[m_sceneLightsIndex] = pLight->GetLightConstant();
-	if (pLight->IsShadowingLight() && m_shadowingLightsIndex < MAX_SHADOWING_LIGHTS_COUNT)
-	{
-		m_shadowingLightsIndices[m_shadowingLightsIndex++] = m_sceneLightsIndex;
-	}
-
-	++m_sceneLightsIndex;
-}
 
 void RRenderer::UpdateGlobalConstant()
 {
