+++ SubModules/Render/RRenderer.cpp
@@ -2,12 +2,35 @@
 #include "pch.h"
 #include "RRenderer.h"
 #include "Core/Components/MeshComponent.h"
+#include "Core/Skybox.h"
+#include "Core/Camera.h"
+#include "Core/Light.h"
+
+
+
+void RRenderer::Render()
+{
+	// Update Scene Resources
+	{
+		UpdateGlobalConstant();
+		// Light
+		m_resourceManager->UpdateStructuredBuffer(sizeof(RLightConstant), static_cast<UINT>(m_sceneLightsIndex), m_sceneLights, m_sceneLightsBuffer);
+	}
+
+	RenderSkybox();
+
+	RenderOpaques();
+
+	RenderTransparent();
+}
+
 
 RMeshGeometry* RRenderer::CreateMeshGeometry(const void* pInVertexList, const UINT vertexSize, const UINT vertexCount, const void* pInIndexList, const UINT indexSize, const UINT indexCount, const EPrimitiveTopologyType topologyType, const EMeshType meshType)
 {
 	MY_ASSERT(pInVertexList != nullptr && pInIndexList != nullptr);
 	return m_resourceManager->CreateMeshGeometry(pInVertexList, vertexSize, vertexCount, pInIndexList, indexSize, indexCount, topologyType, meshType);
 }
+
 
 RMeshGeometry* RRenderer::CreateBasicMeshGeometry(EBasicMeshGeometry type)
 {
@@ -45,6 +68,39 @@ RTexture* RRenderer::CreateTextureCubeFromDDSFile(const WCHAR* wchFileName)
 	MY_ASSERT(wchFileName != nullptr);
 	return m_resourceManager->CreateTextureCubeFromDDSFile(wchFileName);
 }
+RTexture* RRenderer::CreateTexture3D(const UINT width, const UINT height, const UINT depth, const DXGI_FORMAT format)
+{
+	return m_resourceManager->CreateTexture3D(width, height, depth, format);
+}
+
+RTexture* RRenderer::CreateTexture2D(const UINT width, const UINT height, const UINT count, const DXGI_FORMAT format)
+{
+	return m_resourceManager->CreateTexture2D(width, height, count, format);
+}
+
+RTexture* RRenderer::CreateStructuredBuffer(const UINT uElementSize, const UINT uElementCount, const void* pInitData)
+{
+	return m_resourceManager->CreateStructuredBuffer(uElementSize, uElementCount, pInitData);
+}
+
+void RRenderer::SetCamera(const Camera* pCamera)
+{
+	MY_ASSERT(pCamera != nullptr);
+
+	m_camera = pCamera;
+}
+
+void RRenderer::SetSkybox(Skybox* pSkybox)
+{
+	MY_ASSERT(pSkybox != nullptr);
+	m_skybox = pSkybox;
+}
+
+void RRenderer::SetSunLight(const Light* pLight)
+{
+	MY_ASSERT(pLight != nullptr);
+	m_sunLight = pLight;
+}
 
 void RRenderer::SetIBLTextures(const RTexture* pIrradianceMapTexture, const RTexture* pSpecularMapTexture, const RTexture* pBRDFMapTexture)
 {
@@ -59,7 +115,21 @@ void RRenderer::SetIBLTextures(const RTexture* pIrradianceMapTexture, const RTex
 
 }
 
+void RRenderer::AddLight(const Light* pLight)
+{
+	if (m_sceneLightsIndex >= MAX_SCENE_LIGHTS_COUNT)
+	{
+		return;
+	}
 
+	m_sceneLights[m_sceneLightsIndex] = pLight->GetLightConstant();
+	if (pLight->IsShadowingLight() && m_shadowingLightsIndex < MAX_SHADOWING_LIGHTS_COUNT)
+	{
+		m_shadowingLightsIndices[m_shadowingLightsIndex++] = m_sceneLightsIndex;
+	}
+
+	++m_sceneLightsIndex;
+}
 
 void RRenderer::Submit(const MeshComponent* pInMeshComponent, Matrix worldRow)
 {
